% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlot.r
\name{BoxPlot}
\alias{BoxPlot}
\title{Box plot}
\usage{
BoxPlot(File, Projname = NULL)
}
\arguments{
\item{File}{The output of ReadMethylFile function.}

\item{Projname}{A name used to name the plot. The default is null.}
}
\value{
A ggplot2 object
}
\description{
A function to draw a box plot for the DNA methylation dataset.
}
\examples{
data <- Data2[1:10,]
data <- cbind(rownames(data), data)
colnames(data)[1] <- "ID"
BoxPlot(File = data)
}
