% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data1.r
\docType{data}
\name{Data1}
\alias{Data1}
\alias{dataset1}
\title{Training data}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE85212
}
\value{
A data frame
}
\description{
Data1 is a medulloblastoma DNA methylation dataset from a GEO series (GSE85212) and focuses on 399 as the most important probes. This dataset is used to train and test the machine and deep learning models.
}
\examples{
data(Data1)
}
