% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data2.r
\docType{data}
\name{Data2}
\alias{Data2}
\alias{dataset2}
\title{Data2}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE85212
}
\value{
A data frame
}
\description{
Data2 is a medulloblastoma DNA methylation beta values (GSE85212, 50 samples) including 10000 most variable probes used for similarity network fusion.
}
\examples{
data(Data2)
}
\references{
Cavalli FMG, Remke M, Rampasek L, Peacock J et al. Intertumoral Heterogeneity within Medulloblastoma Subgroups. Cancer Cell 2017 Jun 12;31(6):737-754.e6. PMID: 28609654
}
