% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data3.r
\docType{data}
\name{Data3}
\alias{Data3}
\alias{dataset3}
\title{Data3}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE85217
}
\value{
A data frame
}
\description{
Data3 is an gene expression dataset from primary medulloblastoma samples (GSE85217, 50 samples) used for similarity network fusion.
}
\examples{
data(Data3)
}
\references{
Cavalli FMG, Remke M, Rampasek L, Peacock J et al. Intertumoral Heterogeneity within Medulloblastoma Subgroups. Cancer Cell 2017 Jun 12;31(6):737-754.e6. PMID: 28609654
}
