% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.r
\name{ModelMetrics}
\alias{ModelMetrics}
\title{Model metrics}
\usage{
ModelMetrics(Model)
}
\arguments{
\item{Model}{A trained model.}
}
\value{
A data frame
}
\description{
A function to extract the confusion matrix information.
}
\examples{
xgboost <- XGBoostModel(SplitRatio = 0.6,
                        CV = 2,
                        NCores = 1,
                        NewData = NULL)
ModelMetrics(Model = xgboost)
}
