% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewDataPredictionResult.r
\name{NewDataPredictionResult}
\alias{NewDataPredictionResult}
\title{New data prediction result}
\usage{
NewDataPredictionResult(Model)
}
\arguments{
\item{Model}{A trained model}
}
\value{
A data frame
}
\description{
A function to output the predicted medulloblastoma subgroups by trained models.
}
\examples{
set.seed(10)
fac <- ncol(Data1)
NewData <- sample(data.frame(t(Data1[,-fac])),10)
NewData <- cbind(rownames(NewData), NewData)
colnames(NewData)[1] <- "ID"
xgboost <- XGBoostModel(SplitRatio = 0.6,
                        CV = 2,
                        NCores = 1,
                        NewData = NewData)
NewDataPredictionResult(Model = xgboost)
}
