% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadMethylFile.r
\name{ReadMethylFile}
\alias{ReadMethylFile}
\title{Input file for prediction}
\usage{
ReadMethylFile(File)
}
\arguments{
\item{File}{A data frame with tsv or csv file extension. The first column of the data frame is the CpG methylation probe that starts with cg characters and is followed by a number (e.g., cg100091). Other columns are samples with methylation beta values. All columns in the data frame should have a name.}
}
\value{
A data frame
}
\description{
A function to read DNA methylation files. It can be used as the new data for prediction by every model.
}
\examples{
\dontrun{
methyl <- ReadMethylFile(File = "file.csv")
}
}
