% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSNFData.r
\name{ReadSNFData}
\alias{ReadSNFData}
\title{Input file for similarity network fusion (SNF)}
\usage{
ReadSNFData(File)
}
\arguments{
\item{File}{A data frame with tsv or csv file extension. The first column of the data frame is the CpG methylation probe that starts with cg characters and is followed by a number (e.g., cg100091). Other columns are samples with methylation beta values. All columns in the data frame should have a name.}
}
\value{
A data frame
}
\description{
A function to read user-provided file feeding into the SNF function (from the SNFtools package).
}
\examples{
\dontrun{
data <- ReadSNFData(File = "file.csv")
}
}
