% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLabels.r
\docType{data}
\name{RLabels}
\alias{RLabels}
\alias{Subgroups}
\title{RLabels}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE85212
}
\value{
Factor
}
\description{
The actual labels from the medulloblastoma DNA methylation beta values (GSE85212, 50 samples) that was used for similarity network fusion.
}
\examples{
data(RLabels)
}
\references{
Cavalli FMG, Remke M, Rampasek L, Peacock J et al. Intertumoral Heterogeneity within Medulloblastoma Subgroups. Cancer Cell 2017 Jun 12;31(6):737-754.e6. PMID: 28609654
}
