% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSNEPlot.r
\name{TSNEPlot}
\alias{TSNEPlot}
\title{t-SNE 3D plot}
\usage{
TSNEPlot(File, NCluster = 4)
}
\arguments{
\item{File}{The output of ReadMethylFile function.}

\item{NCluster}{The number of cluster.}
}
\value{
Objects of rgl
}
\description{
A function to draw a 3D t-SNE plot for DNA methylation dataset using the K-means clustering technique.
}
\examples{
\donttest{
set.seed(123)
data <- Data2[1:100,]
data <- data.frame(t(data))
data <- cbind(rownames(data), data)
colnames(data)[1] <- "ID"
TSNEPlot(File = data, NCluster = 4)
}
}
