% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadMethylFile.r
\name{ReadMethylFile}
\alias{ReadMethylFile}
\title{Input file for prediction}
\usage{
ReadMethylFile(File)
}
\arguments{
\item{File}{A data frame with tsv or csv file extension. While the first column is CpG methylation probs, starting with cg and followed by a number, other columns are samples with methylation values. All columns should be named.}
}
\value{
A data frame
}
\description{
A function to read DNA methylation files can be used as the new data for prediction by every model.
}
\examples{
\dontrun{
methyl <- ReadMethylFile(File = "file.csv")
}
}
