% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSNFData.r
\name{ReadSNFData}
\alias{ReadSNFData}
\title{Input file for similarity network fusion (SNF)}
\usage{
ReadSNFData(File)
}
\arguments{
\item{File}{A table with tsv or csv file extension. While the first column is a character vector (e.g., gene names or IDs), columns are samples. All columns should be named.}
}
\value{
A data frame
}
\description{
A function to read user-provided file feeding into the SNF function (from the SNFtools package).
}
\examples{
\dontrun{
data <- ReadSNFData(File = "file.csv")
}
}
